Scripts Multiplayer

+++ Assets/Scripts/WordManager.cs
@@ -12,7 +12,7 @@ public class WordManager : MonoBehaviour
     public GUIKeyboardManager guiKeyboardManager;
     public GUIControl guiControl;
 
-    [Header("Text Objects")]
+    [Header("Text Labels")]
     public Text TXTScore;
     public Text TXTTriesCount;
     public Text TXTSuccesfullGuesses;
@@ -21,9 +21,10 @@ public class WordManager : MonoBehaviour
     public string currentGuess = "";
     public List<string> previousWords = new List<string>();
     private int currentScore = 0;
-    private int triesCount = 0;
+    //private int triesCount = 0;
     private int currentTries = 0;
     private int successfulGuesses = 0;
+    public bool isKeyboardLocked = false;
 
     [Header("Colors")]
     public Color defaultColor;
@@ -33,7 +34,7 @@ public class WordManager : MonoBehaviour
     public Color fontOnBlack;
     public Color fontOnWhite;
     
-    private Color[] exits;
+    private Color[] returnedColors;
     private Dictionary<char, Color> letterColorsDict = new Dictionary<char, Color>();
 
     private readonly string validLetters = "qwertyuiopasdfghjklçzxcvbnm";
@@ -49,38 +50,41 @@ public class WordManager : MonoBehaviour
         //    }
         //}
 
-        letterColorsDict.Add('C', correct);
-        letterColorsDict.Add('M', miss);
-        letterColorsDict.Add('I', incorrect);
+        letterColorsDict.Add((char)AttempededLetter.Missed, miss);
+        letterColorsDict.Add((char)AttempededLetter.NotInWord, incorrect);
 
-        exits = new Color[5];
+        returnedColors = new Color[5];
     }
     void Update()
     {
-        if (Input.GetKeyDown(KeyCode.Escape))
+        if (!isKeyboardLocked)
         {
-            guiControl.MenuToggle();
-        }
-        foreach (char letter in Input.inputString.ToLower())
-        {
-            if (letter == '\b')
+			if (Input.GetKeyDown(KeyCode.Escape))
+			{
+                guiControl.MenuToggle();
+                return;
+			}
+            foreach (char letter in Input.inputString.ToLower())
             {
-                DeleteLetter();
-            }
-            else if (letter == '\n' || letter == '\r')
-            {
-                EnterWord();
-            }
-            else if (validLetters.Contains(letter.ToString()))
-            {
-                TypeLetter(letter);
+				
+                if (letter == '\b')
+                {
+                    DeleteLetter();
+                }
+                else if (letter == '\n' || letter == '\r')
+                {
+                    EnterWord();
+                }
+                else if (validLetters.Contains(letter.ToString()))
+                {
+                    TypeLetter(letter);
+                }
             }
         }
     }
 
     public void TypeLetter(char letter)
     {
-        Debug.Log(currentGuess.Length);
         if(currentGuess.Length < 5)
         {
             currentGuess += letter;
@@ -132,14 +136,24 @@ public class WordManager : MonoBehaviour
     }
 
     /// <summary>
-    /// Envia tentativa do usuário para wordsManager e verifica se o resultado recebido é uma palavra acertada.
+    /// Envia tentativa do usuário para wordsManager; para o singleplayer, o próximo passo é invocar ReceiveAttemptedLetters().
     /// </summary>
     private void ReviseLetters()
     {
-        bool allCorrect = true;
-        Debug.Log(currentGuess);
+        isKeyboardLocked = true;
+        Debug.Log($"ReviseLetters - {currentGuess}");
         string checkedWord = wordsManager.GetCheckedAttempt(currentGuess);
-        Debug.Log("attempt string chain: " + new string(checkedWord));
+        if(checkedWord != "") ReceiveAttemptedLetters(checkedWord);
+    }
+
+    /// <summary>
+    /// Verifica se o resultado recebido é uma palavra acertada e gera a lógica de pintar teclas/letras.
+    /// </summary>
+    public void ReceiveAttemptedLetters(string checkedWord)
+    {
+        Debug.Log($"Recebendo {checkedWord}");
+        Debug.Log($"currentGuess: {currentGuess}");
+        bool allCorrect = true;
         if (checkedWord == "X")
         {
             ShakeLetters();
@@ -147,31 +161,39 @@ public class WordManager : MonoBehaviour
         }
         else
         {
+            System.Text.StringBuilder checkedWordEditable = new System.Text.StringBuilder(checkedWord);
             for (int i = 0; i < checkedWord.Length; i++)
             {
-                exits[i] = letterColorsDict[checkedWord[i]];
-                allCorrect = allCorrect && (checkedWord[i] == 'C');
+                try
+                {
+                    returnedColors[i] = letterColorsDict[checkedWord[i]];
+                    checkedWordEditable[i] = currentGuess[i];
+                    allCorrect = false;
+                }
+                catch
+                {
+                    returnedColors[i] = correct;
+                }
+                //allCorrect = allCorrect && (checkedWord[i] != (char)AttempededLetter.Missed && checkedWord[i] != (char)AttempededLetter.NotInWord);
             }
 
+            currentTries++;
             if (allCorrect)
             {
-                previousWords.Add(currentGuess);
+                previousWords.Add(checkedWord);
                 IncreaseScore();
-            }
-            else
-            {
-                triesCount++;
-                currentTries++;
             }
 
             string currentGuessNoAccents = SinglePlayerTextManipulation.RemoveAccents(currentGuess);
-            guiAnswerManager.CreateAnswerSprites(allCorrect ? previousWords[previousWords.Count - 1] : currentGuessNoAccents, exits, fontOnWhite);
-            guiKeyboardManager.PaintKeyboardLetters(currentGuessNoAccents.ToUpper(), exits);
+            //guiAnswerManager.CreateAnswerSprites(allCorrect ? checkedWord : currentGuessNoAccents, returnedColors, fontOnWhite);
+            guiAnswerManager.CreateAnswerSprites(checkedWordEditable.ToString(), returnedColors, fontOnWhite);
+            guiKeyboardManager.PaintKeyboardLetters(currentGuessNoAccents.ToUpper(), returnedColors);
 
             UpdateStatsText();
             ClearGuess(isAllCorrect: allCorrect);
             currentGuess = "";
         }
+        isKeyboardLocked = false;
     }
 
     private void ClearGuess(bool isAllCorrect)
@@ -206,14 +228,14 @@ public class WordManager : MonoBehaviour
     private void IncreaseScore()
     {
         successfulGuesses++;
-        currentScore += currentTries <= 0 ? 1000 : 1000/currentTries;
+        currentScore += currentTries <= 0 ? (currentTries > 5 ? 0 : 1000) : 1000/currentTries;
         currentTries = 0;
     }
 
     private void UpdateStatsText()
     {
         TXTScore.text = "Pontos: " + currentScore.ToString();
-        TXTTriesCount.text = "Tentativas: " + triesCount.ToString();
+        TXTTriesCount.text = "Tentativas: " + currentTries.ToString();
         TXTSuccesfullGuesses.text = "Acertos: " + successfulGuesses.ToString();
     }
 }
